HaF save 'correct' status to store for practice and test stimulus trials
TODO: double-check what needs to be saved to store and to jsPsych.data.addDataToLastTrial
