--- task-launcher/src/tasks/hearts-and-flowers/trials/enums.js
@@ -1,20 +1,0 @@
-
-export const StimulusType = Object.freeze({
-  Heart: 'heart',
-  Flower: 'flower',
-});
-
-// Enum for the side the stimulus is shown on and also for button_choices
-export const StimulusSideType = Object.freeze({
-  Left: 'left',
-  Right: 'right',
-});
-
-export const AssessmentStageType = Object.freeze({
-    HeartsPractice: 'hearts practice',
-    FlowersPractice: 'flowers practice',
-    HeartsStimulus: 'hearts stimulus',
-    FlowersStimulus: 'flowers stimulus',
-    HeartsAndFlowersPractice: 'hearts and flowers practice',
-    HeartsAndFlowersStimulus: 'hearts and flowers stimulus',
-});
